console.py: split hardware manager
